fix build modes for windows
