+++ pkg/network/tracer/tracer_windows_test.go
@@ -10,10 +10,11 @@ package tracer
 import (
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/network/testutil"
+	"github.com/stretchr/testify/suite"
 
 	"github.com/DataDog/datadog-agent/pkg/network/config"
 	"github.com/DataDog/datadog-agent/pkg/network/driver"
+	"github.com/DataDog/datadog-agent/pkg/network/testutil"
 )
 
 func platformInit() {
@@ -27,6 +28,10 @@ func httpSupported() bool {
 func testConfig() *config.Config {
 	cfg := config.New()
 	return cfg
+}
+
+func TestTracerSuite(t *testing.T) {
+	suite.Run(t, new(TracerSuite))
 }
 
 // nolint:unused   // this function currently unused but will be.
